+++ src/components/layout/Layout/TopBanner.tsx
@@ -15,9 +15,8 @@ const TopBanner = () => {
         fontSize: '14px',
       }}
     >
-      <strong>AR.IO is migrating to Solana!</strong> ArNS name purchases are
-      paused and will resume after launch. All names will be snapshotted May 15,
-      2026.{' '}
+      <strong>Ar.io is migrating to Solana!</strong> Purchases are paused and
+      will resume shortly. Register before the May 15, 2026 snapshot!{' '}
       <Link
         to={SOLANA_MIGRATION_LINK}
         target="_blank"